Use our own PRNG
We can't rely on drand48() because it's seeded, so after forking a process
both parent and child will use the same random numbers. That makes it
unusable for some testing scenarios.

In this patch, we implement our own PRNG based on a linear congruential
generator, the same one used by rand() and drand48(), but altering it to
take the timestamp into account.

While it breaks some of the generator properties, it works well enough for
our purposes, and we get rid of the issues with fork().

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>