Add a new preload library to enable failure points before a program starts
This patch adds a new preload library which allows the user to enable
failure points and remote control (via environment variables) before a
program starts.

It can be used combined with the posix preload to simulate faults in the
standard library functions.

A shell script to make it easier to use will be added in future commits,
along more documentation.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>