preload: Do not use constructor priorities when they're not supported
Constructor priorities are supported only from gcc 4.3 on, and only on some
platforms.

Since we don't rely on them, this patch avoids using them when building on
an unsupported system.

Note that clang defines __GNUC__, so when it gets constructor priorities
support, it should work fine without needing a special case.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>