preload/posix: Simulate incomplete reads and writes
This patch implements a generic way of adding a point of failure to reduce
a given parameter in a random amount, which is then used to add points of
failure to simulate incomplete reads and writes.

Note that it will never make it 0, as that usually has a special meaning
(EOF for reads).

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>