Escape "-" to ensure interpretation as minus sign, not hyphen
"-" chars are interpreted as hyphens (U+2010) by groff, not as minus
signs (U+002D). Since options to programs use minus signs (U+002D), this
means for example in UTF-8 locales that you cannot cut&paste options,
nor search for them easily.

"-" must be escaped ("\-") to be interpreted as minus. If we really
intend a hyphen, we can write it as "\(hy" to emphasise that fact. See
groff(7) and especially groff_char(7) for details.

It might be possible to fix ths by jumping into a different context,
but I don't know enough manpage stuff.

Signed-off-by: Chris Lamb <lamby@debian.org>