Don't assume that we will use the library from the 'install' target's PREFIX
This patch adds support for using the DESTDIR variable in the 'install'
target to represent the "eventual" destination of the library, decoupling
the installation directory and the run-time directory.

I'm using this so the pkgconfig file gets created with a "prefix" value
of my choice ("/usr") instead of somewhere under my home directory or
wherever I happen to be generating the Debian package from.

Signed-off-by: Chris Lamb <lamby@debian.org>