preload/posix: Fix initialization when building without constr. priorities
When building without constructor priorities, the order in which they are
called is not specified.

This caused several strange bugs (segfaults and hangs) mostly because
three issues:

 - Wrappers wanted to call the original libc version, but _fiu_init()
   hadn't performed the dlopen() yet.
 - _fiu_init() got called with _fiu_called > 0, but it resetted it back
   to 0, so the reference counter was left wrong.
 - Wrappers were called before their own constructor (i.e. from another
   constructor).

This patch fixes those things by making the wrapper constructors to call
_fiu_init() by hand if _fiu_libc == NULL, making _fiu_init() cope with
that, and fixing the wrappers to call their own constructors if they
haven't been called yet.

Thanks to Stas Oskin for the detailed bug report.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>