preload/posix: Pass an appropriate promoted type to va_arg()
va_arg() can only take a promoted type, which that means just an int in this
case.

On some platforms passing mode_t works, but on others (like Debian's kFreeBSD
amd64) causes a compile-time warning and a run-time error, as reported by
Serafeim Zanikolas on Debian bug 574752.

This patch fixes this by passing int as the argument to va_arg(), instead of
mode_t.

It also fixes the "mode" variable's type, which was int when it should have
been mode_t.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>