preload: Make detection of libc's soname and -ldl more robust
This patch turns the detection of libc's soname and -ldl requirement from a
make macro into a target, which is slightly more robust.

It also avoids invoking the linker directly, but uses the compiler instead.
This is used to avoid problems when the way we invoke the linker is not the
same way the compiler does, which can cause troubles.

Specifically, on Fedora 14 Beta, running ld by hand did not pass the
requirement of a non-executable stack, which broke ldd's processing of the
resulting shared library (because of selinux).

It's still not the nicest thing to do, but it's better than using autoconf.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>