preload/posix: Use limits.h instead of features.h for portability reasons
We want to know if we're building against glibc. glibc defines __GLIBC__ for
those purposes, and to get it defined we must include features.h.

The problem is features.h is NOT standard, and not available on some
platforms.

So we include limits.h, because we know it includes features.h, and we get the
define indirectly.

This is UGLY but sadly the other choices are not much more attractive. This
may get better if we used some script like autotools to detect features.h
availability, and we may do that in the future.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>