libfiu: Move the method-specific setup out of the common setup_fail()
Currently, the internal function setup_fail() takes all posible parameters to
set up a failure point, and knows what to set for each method.

That results in an ugly API, because it has to take many unnecessary
parameters, and the code is harder to follow and more complicated.

This patch removes that knowledge from setup_fail(), which now just sets up
the basic fields, and leaves the specific fields for each of the
fiu_enable_*() functions.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>