libfiu: Implement fiu_enable_stack() and fiu_enable_stack_by_name()
This patch implements two new ways of enable points of failure:
fiu_enable_stack() and fiu_enable_stack_by_name().

It allows the user to say "enable the point of failure X, if the function F is
on the stack".

When X is evaluated, it traverses the stack looking for F, and makes the point
fail if the function is found. This relies on some GNU-specific functions, so
it's not portable, but it should fail gracefully on platforms where it's not
available.

There's support in the API for specifying the position of the function F,
although it has not been implemented yet.

Some very basic tests are introduced, more tests will probably be added in
later patches.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>