Simplify remote control commands
The remote control commands were specified in many different places, with
different ad-hoc parsing and passing to get them from the user to the remote
control thread.

This patch rewrites how remote control commands get parsed and passed around,
unifiying the parsing in the library itself, with a much more extensible and
reasonable format.

The command line tools are also updated to get the commands directly from the
user, and passing them without any need for parsing. The older flags for
specifying commands are marked as deprecated and will be removed in future
releases.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>