preload/posix: Use RTLD_NEXT when available
This patch makes the posix preloader use RTLD_NEXT for finding the original
symbols of the functions we're wrapping, instead of needing
dlopen(LIBC_SONAME). It's a GNU extension, but very very useful.

The change also requires some reorganization, including a special case for
failing the function during its own initialization, which can happen if
dlopen()/dlsym() invoke the functions we are wrapping.

That works as long as those functions can cope with the failures, which is the
case in all tested platforms so far (and we're no worse than we were before
anyway).

Much of the initialization and libc handling is simplified, so the patch
removes more than twice the lines it adds.

The original motivation for this patch, however, is to fix the removal of the
malloc hooks in the latest glibc versions. With it, malloc() and friends no
longer need special treatment. Thanks a lot to Georg Profos for reporting this
issue, along with code samples for reproducing it.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>