tests: Do not fail if fiu_enable_stack*() returns an error
fiu_enable_stack() and fiu_enable_stack_by_name() are only supported on some
platforms; where they're not, an error will be returned.

This small patch avoid failing the tests on that situation. Is not ideal
because the function could fail when it shouldn't, but until we introduce
other changes needed for detecting when it should and should not work, this
prevents the false negative.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>