libfiu: Test if backtrace works in fiu_enable_stack()
In fiu_enable_stack(), test the backtrace-related functions to make sure they
work before adding the new failure point.

That way, the caller will know if the function isn't available on the platform
instead of the failure point not working when it should.

We need it because even on platforms where backtrace() is available, it may
not be working properly (because of build options, broken glibc, etc.).

The tester function takes the caller as a parameter to work around compiler
inlining issues that can result in a false negative.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>