libfiu: Fix backtrace_works() in fiu_enable_stack_by_name()
fiu_enable_stack_by_name() calls fiu_enable_stack(), which does the
backtrace_works() check.

For that check to work, fiu_enable_stack() must be on the stack. However, when
fiu_enable_stack_by_name() invokes it, it could be inlined, causing the check
to fail.

This patch adds an independent backtrace_works() check within
fiu_enable_stack_by_name(), so we're sure the caller is on the stack.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>