libfiu: Use a separate data structure to track enabled failure points
This patch introduces a new simple data structure, the wtable, which is used
to store key-value pairs (with \0-terminated keys and void * values), that
supports wildcarded keys.

It is then used to replace the open-coded enabled_fails structure in fiu.c.

Incidentally, that fixes a couple of races at insertion time (e.g. in
fiu_enable_external() we could attempt to call the external function pointer
before we had properly set it in our pf_info structure).

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>