preload/posix: add a way to generate 64-bit counterparts
Some functions (for example, pread()) have their 64-bit counterparts for
the case when off_t is 64-bit on 32-bit system (e.g. FILE_OFFSET_BITS=64
macro is defined). If the macro is defined, a binary will be linked with
<func>64() symbol instead of <func>() and libfiu won't be able to hook
it.

This might not be that important since 32-bit systems are rare now,
but, despite the fact that feature_test_macros(7) man states that there
is no effect on 64-bit system, there actually is. Even on 64-bit systems
when FILE_OFFSET_BITS=64 is defined, <func>64() symbols is used instead
of <func>().

Since it is a good practice to actually define this macro, libfiu can't
hook any of such functions.

To fix this, add a special keyword "variant: off64_t" to function
description and output two variants of a function if such a keyword is
present.

Signed-off-by: Pavel Borzenkov <pavel@voidptr.ru>