Do not use the dummy backtrace on profiling builds
When building with PROFILE=1, the current USE_DUMMY_BACKTRACE test fails
because the compiler tries to write to /dev/null.gcno.

This patch fixes it by making it write to a local file instead, which we then
not use.

It also adds the profiling files to .gitignore while at it.