tests: Make random tests more robust
Some of the tests which check the random operations depend on /bin/cat doing
I/O operations a certain way, which is not portable and not very accurate
either. This results in these tests occasionally failing without good reason.

To fix that, we introduce a small "cat" implementation with very simple and
predictable I/O, which we can rely upon in the tests.