libfiu: Filter the list of exported symbols
The library currently exports some symbols, like hash_create, which are not
meant to be exported and can cause collisions with other libraries.

This patch introduces a list of explicitly exported symbols used by the linker
to generate the shared library accordingly.

Thanks to Tim Cuthbertson for reporting the problem and sending a patch.