Move libraries after files that need them to fix FTBFS with ld --as-needed.
[ Alberto's notes follow ]

FTBFS means "Fails To Build From Source", https://wiki.debian.org/FTBFS.

This patch fixes the build when "--as-needed" is passed to the linker.
The problem can be reproduced by running:

  make CFLAGS=-Wl,--as-needed clean test

The --as-needed linker option is used when building packages for some
distributions, notably Debian and Ubuntu.

The patch was written by Logan Rosen, who applied the patch to Ubuntu's
package, the original can be found at:
http://bazaar.launchpad.net/~ubuntu-branches/ubuntu/wily/libfiu/wily/revision/20

Chris Lamb applied it to Debian's package, and told me about it so I
could merge it directly (which I'm doing now :).

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=819373
https://github.com/lamby/pkg-libfiu/commit/f8cc48a0f5cc45454ac5ef3d59725e751abe3833

Thanks a lot to both!