tests/utils: Set LC_ALL=C for tests that are locale-dependant
The test-basic_ctrl.py test is locale dependant, as it checks the output
of cat for the word "error", which only works on certain locales.

This patch fixes the locale to "C" which should be more reliable and
reproducible, and while at it increase the logging in case of errors so
these kind of problems are easy to detect.

Thanks to Chris Lamb for finding and reporting the problem.