tests/utils: Fix broken dependency chain in Makefile
There is a bug in tests/utils/Makefile, where the tests themselves don't
depend on the target for their libraries.  The higher-level target "all"
(the default) does, and that's why it works under most circumstances.

However, a parallel build can fail because of this, as the ordering is
not guaranteed. For example, this causes it to break on my workstation:

  make clean; ( cd tests/utils; make clean); make; V=1 make -j 8 tests

This bug was found by the Debian build system (thanks!), see
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=893049 for more
details.

So this patch fixes the problem by making the tests depend on the
appropriate targets.