Add a function to manually set the PRNG seed (fiu_set_prng_seed())
When doing randomized tests, it can be useful to have a way of
reliably reproducing failures.  Currently this is not possible, because
the PRNG seed is set internally based on time, and users have no control
over it.

So this patch introduces a new function, fiu_set_prng_seed(), to allow
users to manually control the PRNG seed.