preload/posix: Use macros to generate wrappers for variadic functions
To wrap some of the variadic functions like fprintf(), we can rely on a
non-variadic counterpart like vprintf().

Today this is implemented by open-coding the wrapper, causing a
significant amount of duplication (because they're all very similar to
each other, and they're all very similar to the codegen.h macros).

To reduce that duplication, this patch adds some specific macros to
generate these variadic wrappers.