preload/posix: Replace open-coded hash table with the existing one
In preload/posix/modules/posix.custom.c we need to keep a set of FILE *.
Current implementation open-codes a hash table of fixed size.

This patch replaces that with an existing hash table implementation.
The existing one is dynamic in size, but requires to convert the
pointers to string keys, so it has a performance penalty.

This change is mostly to make the code easier to maintain, as ~halves
the lines of code, and reuses a more tested hash implementation.