tests: Add test for the handling of FILE * errors
The handling of FILE * errors is tricky, as we keep track of the streams
with errors internally.

This patch adds a test that exercises some of the cases related to error
handling. It is not overly tricky, but it covers quite a few corner
cases and already detected a couple of problems (fixed in previous
patches).