preload/posix: Enforce initialization of ferror hash table
The hash table we use to keep track of ferrors is currently initialized
via a constructor.

If for some reason the functions using the table are run before the
constructor has had a chance to run (for example, early init of other
libraries), then the library crashes because it tries to operate on a
NULL hash table.

This patch fixes this by always checking that the table has been
initialized before using it, and it does so in a lazy way (purely for
convenience).

Thanks to Pietro Oliva for finding this issue, and the detailed bug
report!