tests/generated: fprintf() test should successfuly flush
Gianfranco Costamagna (locutusofborg@debian.org) reported that the
automatic builds in Ubuntu have been failing on the fprintf() generated
tests, on all platforms.  The tests pass the successful case, but fail
to invoke the callback.

This is very suspicious. Gianfranco had a theory about /dev/null writes,
and while looking into it, I noticed that the libc buffering might be
making the success case pass even if the file can't be written to.

So this patch forces the success case to do a flush, to try to identify
these situations early.