bindings/python: Increment refcount on failinfo() objects
When we return objects from failinfo(), we need to increment their
refcount so that they are properly tracked and accounted for. Otherwise,
the GC might unreference them to early.

This was always present but changes in Python 3.8 makes it easier to
trigger, and was causing a test failure caught by Debian tests:
https://bugs.debian.org/954287