bindings/python: Remove unnecessary parameter in failinfo()
The failinfo() Python function takes a name as parameter, but it is
misleading as it is never used, and the returned value has no connection
to it.

This is probably an ancient copy-paste error that has gone unnoticed.

This is a breaking API change. However, given how rare calls to this
function are, and that it is probably unused in the wild, and that it
is trivial to fix, I'm going to just mention it in the release notes but
don't think it merits a new major version.