fiu-run: Avoid uninitialized memory on error in fiu_run_init()
In fiu_rc_string(), when a command is valid but returns an error (for
example, disabling a name that has not been enabled), no error message
is returned.

In turn, that causes an access to non-initialized memory in
fiu_run_init() when printing the error message.

This patches fixes the issue by always initializing the error message.

Amended-by: Alberto Bertogli <albertito@blitiri.com.ar>
  Edited commit message.