Deprecate Python 2 usage
Python 2 is end-of-life.

We don't use Python 2 for building or testing, but currently have Python
2 bindings that get built with "make all" and "make test".

This patch removes the Python 2 make target dependency, so Python 2
bindings won't get automatically built anymore.

Manual builds continue to be supported (for now) with "make python2" and
friends, but are likely to be removed in the future.