tests: Prevent compiler from inlining stack test functions
The tests for enable_stack() and enable_stack_by_name() depend on being
able to identify specific functions in the stack. If those functions are
inlined, they may not appear in the backtrace information and cause the
tests to fail.

For example, this happens when the test is built with
CFLAGS="-flto=auto -O2", as reported in https://bugs.debian.org/1015487.

This patch fixes the problem by telling the compiler not to inline those
specific test functions. The attribute used works at least with GCC
(tested), clang (tested), and icc (based on documentation).

The documentation is also updated to explicitly mention this, to prevent
users from getting confused due to this same problem on their own builds.