tests: Skip fprintf() test if it can't be wrapped
Under some conditions, fprintf() is declared as a macro. And when
_FORTIFY_SOURCE is set, it can be declared in the headers and forced to
be inlined. In both cases, the preload can't work because the function
is never looked up outside the binary.

This has caused some test failures in the Ubuntu packaging automated
builds, which builds the libfiu packages using fortified flags.

Thanks to Gianfranco Costamagna <locutusofborg@debian.org> for
identifying and reporting this issue.