tests: Update wrap-python to support newer distutil versions
The `wrap-python` script tries to detect the build location for the
Python bindings, but unfortunately newer distutil versions have changed
the directory layout and it is no longer finding them.

This detection method is brittle in general, but it was needed to
support multiple Python major and minor versions.

However, these days where only Python 3 is supported, we can probably
remove it and rely on the build being consistent, or at least ask the
user for a "make clean" between minor Python versions.

This is a minimal patch that gets things functional again by loosening
the glob to find the build directory.

In future patches, we may remove this mechanism altogether and simplify
wrap-python.