fiu-run: Fix shell error in deprecated option parsing
When using the deprecated options (e.g. `-e`), there is a bug where some
of the bash comparisons are invalid because the variable names are
incorrect.

This bug has been present since these flags were deprecated in 2012
(commit c5f4c471).

Thanks to Lynn Kerby <lfkerby@gmail.com> who reported this issue
and provided a patch for it!

Co-authored-by: Lynn Kerby <lfkerby@gmail.com>