preload/posix: Specify void in empty function declarations
Some functions (like `fork()`) have no parameters. In such cases, the code
generator would produce an empty PARAMST.

When those are compiled with `-Wstrict-prototypes`, that will result in
warnings on the `_fiu_init_##NAME()` functions.

To fix this, we explicitly set `void` on those cases, as per modern C
standards.

https://github.com/albertito/libfiu/pull/11

Amended-by: Alberto Bertogli <albertito@blitiri.com.ar>
  Edited commit message, added comment.