tests/generated: Disable optimizations to avoid false positives
The code of the generated tests are fairly simple, and the compiler may
inline or optimize away some of the calls of the functions under test,
causing false positives.

Unfortunately, there's nothing we can do about intercepting inlined
functions (and much less calls that are optimized away), that's an
intrinsic limitation of the library and nothing new. And as compilers
change, the optimizations may cause different false positives in our
tests.

This happened in https://bugs.debian.org/1097184 where the malloc() call
was being optimized away by `gcc-15 -O2`, and causing a new false
positive.

This patch fixes/works around the problem by passing `-O0` to disable
optimizations in the compilation of the generated tests.

Ideally we would do something more targeted, but for now this is good
enough to prevent most of the problems.